+++ lib/rblade/component_store.rb
@@ -14,7 +14,7 @@ module RBlade
     def component(full_name)
       # If this is a relative path, prepend with the previous component name's base
       if full_name.start_with? "."
-        full_name = @component_name_stack.last.gsub(/\.[^\.]++\z/, "") + full_name
+        full_name = @component_name_stack.last.sub(/[^.]++\z/, "") + full_name.delete_prefix(".")
       end
 
       # Ensure each component is only compiled once
@@ -24,14 +24,7 @@ module RBlade
 
       @component_name_stack << full_name
 
-      namespace = nil
-      name = full_name
-
-      if name.match? "::"
-        namespace, name = full_name.split("::")
-      end
-
-      method_name = compile_component full_name, File.read(find_component_file(namespace, name))
+      method_name = compile_component full_name, File.read(find_component_file(full_name, @component_name_stack))
       @component_name_stack.pop
 
       method_name
@@ -55,9 +48,11 @@ module RBlade
       @component_definitions
     end
 
-    private
+    def self.find_component_file(name, name_stack = nil)
+      if name.match? "::"
+        namespace, name = name.split("::", 2)
+      end
 
-    def find_component_file(namespace, name)
       file_path = name.tr ".", "/"
 
       template_paths[namespace]&.each do |base_path|
@@ -67,7 +62,7 @@ module RBlade
           end
           if File.exist? base_path + file_path + "/index" + extension
             # Account for index files for relative components
-            @component_name_stack << @component_name_stack.pop + ".index"
+            name_stack << name_stack.pop + ".index" if name_stack
             return "#{base_path}#{file_path}/index#{extension}"
           end
         end
@@ -75,6 +70,9 @@ module RBlade
 
       raise RBladeTemplateError.new "Unknown component #{namespace}::#{name}"
     end
+    delegate :find_component_file, to: :class
+
+    private
 
     def compile_component(name, template)
       escaped_name = name.gsub(/[^0-9a-zA-Z_]/) do |match|
@@ -84,14 +82,10 @@ module RBlade
 
       compiled_component = RBlade::Compiler.compileString(template, self)
 
-      slot_assignment = compiled_component.match?(/\Wslot\W/) ? "slot=" : ""
-
-      @component_definitions << "def self._rblade_component_#{escaped_name}(attributes,&);#{slot_assignment}if block_given?;RBlade::SlotManager.new(@output_buffer.capture(->(name, slot_attributes, &slot_block)do;attributes[name]=RBlade::SlotManager.new(@output_buffer.capture(&slot_block), slot_attributes);end,&));end;_stacks=[];@output_buffer.raw_buffer<<@output_buffer.capture do;#{compiled_component}@output_buffer.raw_buffer.prepend(@_rblade_stack_manager.get(_stacks));end;end;"
+      @component_definitions << "def self._rblade_component_#{escaped_name}(attributes,&);slot=if block_given?;RBlade::SlotManager.new(@output_buffer.capture(->(name, slot_attributes, &slot_block)do;attributes[name]=RBlade::SlotManager.new(@output_buffer.capture(&slot_block), slot_attributes);end,&));end;_stacks=[];@output_buffer.raw_buffer<<@output_buffer.capture do;#{compiled_component}@output_buffer.raw_buffer.prepend(@_rblade_stack_manager.get(_stacks));end;end;"
 
       @component_method_names[name] = "_rblade_component_#{escaped_name}"
     end
-
-    private
 
     cattr_accessor :template_paths, default: {}
   end